chore: rename package to apps
