+++ core/domain/src/main/kotlin/dev/mslalith/focuslauncher/core/domain/apps/GetAppsUseCase.kt
@@ -1,4 +1,4 @@
-package dev.mslalith.focuslauncher.core.domain.appswithicons
+package dev.mslalith.focuslauncher.core.domain.apps
 
 import dev.mslalith.focuslauncher.core.domain.extensions.toAppWithIcons
 import dev.mslalith.focuslauncher.core.domain.extensions.toAppsWithColor
