+++ core/domain/src/main/kotlin/dev/mslalith/focuslauncher/core/domain/apps/GetIconPackAppsWithIconsUseCase.kt
@@ -1,4 +1,4 @@
-package dev.mslalith.focuslauncher.core.domain.appswithicons
+package dev.mslalith.focuslauncher.core.domain.apps
 
 import dev.mslalith.focuslauncher.core.domain.iconpack.GetIconPackAppsUseCase
 import dev.mslalith.focuslauncher.core.model.app.AppWithIcon
