+++ feature/appdrawerpage/src/main/kotlin/dev/mslalith/focuslauncher/feature/appdrawerpage/AppDrawerPageViewModel.kt
@@ -8,7 +8,7 @@ import dev.mslalith.focuslauncher.core.data.repository.AppDrawerRepo
 import dev.mslalith.focuslauncher.core.data.repository.FavoritesRepo
 import dev.mslalith.focuslauncher.core.data.repository.HiddenAppsRepo
 import dev.mslalith.focuslauncher.core.data.repository.settings.AppDrawerSettingsRepo
-import dev.mslalith.focuslauncher.core.domain.appswithicons.GetAppDrawerAppsWithIconsUseCase
+import dev.mslalith.focuslauncher.core.domain.apps.GetAppDrawerAppsWithIconsUseCase
 import dev.mslalith.focuslauncher.core.model.Constants.Defaults.Settings.AppDrawer.DEFAULT_APP_DRAWER_VIEW_TYPE
 import dev.mslalith.focuslauncher.core.model.Constants.Defaults.Settings.AppDrawer.DEFAULT_APP_GROUP_HEADER
 import dev.mslalith.focuslauncher.core.model.Constants.Defaults.Settings.AppDrawer.DEFAULT_APP_ICONS
